+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.2
+### Added
+- Implement string similarity UDFs and Splink function name hooks:
+  - Jaccard Similarity function name (using undocumented Snowflake function)
+  - Jaro Similarity UDF
+  - Damerau-Levenshtein UDF
+- Add UDF tests
+
 ## 0.1.1
 ### Added
 - Add OIDC / Workload Identity authentication support for GitHub Actions CI